+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -63,6 +63,8 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
             "IsCaptureStateWidgetPlugin",
             "IsCategoryHiddenStateWidgetPlugin",
             "IsVisibleActionWidgetPlugin",
+            "NicknameToggleActionWidgetPlugin",
+            "PrimRenameNameActionWidgetPlugin",
             "PrimTreeWidgetPlugin",
         ],
         exclude=True,