+++ source/extensions/omni.flux.stage_manager.plugin.listener.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.3.0]
+## Added
+- Added nickname attribute changes to the listener and only refresh if there are new paths changes
+
 ## [2.2.0]
 ## Added
 - Added a `cleanup` pattern to the stage manager context and listener plugins